+++ lib/src/main/java/com/wjduquette/joe/bert/TokenType.java
@@ -6,6 +6,7 @@ package com.wjduquette.joe.bert;
 enum TokenType {
     // Single-character tokens.
     LEFT_PAREN, RIGHT_PAREN, LEFT_BRACE, RIGHT_BRACE,
+    LEFT_BRACKET, RIGHT_BRACKET,
     AT, BACK_SLASH, COLON, COMMA, DOT, QUESTION,
     SEMICOLON,