+++ lib/src/main/java/com/wjduquette/joe/bert/VirtualMachine.java
@@ -2,6 +2,7 @@ package com.wjduquette.joe.bert;
 
 import com.wjduquette.joe.*;
 import com.wjduquette.joe.types.ListValue;
+import com.wjduquette.joe.types.MapValue;
 
 import java.util.*;
 
@@ -308,7 +309,7 @@ class VirtualMachine {
                 case CONST -> push(readConstant());
                 case DECR -> {
                     var a = pop();
-                    checkNumericOperand("--", a);
+                    checkPrePostOperand("--", a);
                     push((double)a - 1);
                 }
                 case DIV -> {
@@ -318,6 +319,10 @@ class VirtualMachine {
                     push((double)a / (double)b);
                 }
                 case DUP -> push(peek(0));
+                case DUP2 -> {
+                    push(peek(1));
+                    push(peek(1));
+                }
                 case EQ -> {
                     var b = pop();
                     var a = pop();
@@ -392,8 +397,40 @@ class VirtualMachine {
                 }
                 case INCR -> {
                     var a = pop();
-                    checkNumericOperand("++", a);
+                    checkPrePostOperand("++", a);
                     push((double)a + 1);
+                }
+                case INDGET -> {
+                    var index = pop();
+                    var coll = pop();
+
+                    switch (coll) {
+                        case JoeList list -> {
+                            var i = checkListIndex(list, index);
+                            push(list.get(i));
+                        }
+                        case JoeMap map -> push(map.get(index));
+                        default ->
+                            throw error("Expected indexed collection, got: " +
+                                joe.typedValue(coll) + ".");
+                    }
+                }
+                case INDSET -> {
+                    var value = pop();
+                    var index = pop();
+                    var coll = pop();
+
+                    switch (coll) {
+                        case JoeList list -> {
+                            var i = checkListIndex(list, index);
+                            list.set(i, value);
+                        }
+                        case JoeMap map -> map.put(index, value);
+                        default ->
+                            throw error("Expected indexed collection, got: " +
+                                joe.typedValue(coll) + ".");
+                    }
+                    push(value);
                 }
                 case INHERIT -> {
                     var superclass = peek(1);
@@ -450,6 +487,12 @@ class VirtualMachine {
                         throw error("The '<=' operator expects two Numbers or two Strings.");
                     }
                 }
+                case LISTADD -> {
+                    var item = pop();
+                    var list = (ListValue)peek(0);
+                    list.add(item);
+                }
+                case LISTNEW -> push(new ListValue());
                 case LOCGET -> {
                     var slot = readSlot();
                     push(stack[frame.base + slot]);
@@ -472,6 +515,13 @@ class VirtualMachine {
                     } else {
                         throw error("The '<' operator expects two Numbers or two Strings.");
                     }
+                }
+                case MAPNEW -> push(new MapValue());
+                case MAPPUT -> {
+                    var value = pop();
+                    var key = pop();
+                    var map = (MapValue)peek(0);
+                    map.put(key, value);
                 }
                 case METHOD -> {
                     // NOTE: This was defineMethod in clox
@@ -673,9 +723,24 @@ class VirtualMachine {
         }
     }
 
-    private void checkNumericOperand(String op, Object a) {
+    private void checkPrePostOperand(String op, Object a) {
         if (!(a instanceof Double)) {
-            throw error("The '" + op + "' operator expects a numeric operand.");
+            throw error("Target of '" + op + "' must contain a number.");
+        }
+    }
+
+    private int checkListIndex(List<?> list, Object index) {
+        if (index instanceof Double d) {
+            int i = d.intValue();
+            if (i >= 0 && i < list.size()) {
+                return i;
+            } else {
+                throw error("List index out of range [0, " +
+                    (list.size() - 1) + "]: " + i + ".");
+            }
+        } else {
+            throw error("Expected list index, got: " +
+                joe.typedValue(index) + ".");
         }
     }