+++ lib/src/main/java/com/wjduquette/joe/walker/Resolver.java
@@ -220,14 +220,29 @@ class Resolver {
             }
             case Expr.Get expr -> resolve(expr.object());
             case Expr.Grouping expr -> resolve(expr.expr());
+            case Expr.IndexGet expr -> {
+                resolve(expr.collection());
+                resolve(expr.index());
+            }
+            case Expr.IndexSet expr -> {
+                resolve(expr.collection());
+                resolve(expr.index());
+                resolve(expr.value());
+            }
             case Expr.Lambda expr ->
                 resolveFunction(expr.declaration(), FunctionType.LAMBDA);
+            case Expr.ListLiteral expr -> expr.list().forEach(this::resolve);
             case Expr.Literal ignored -> {}
             case Expr.Logical expr -> {
                 resolve(expr.left());
                 resolve(expr.right());
             }
+            case Expr.MapLiteral expr -> expr.entries().forEach(this::resolve);
             case Expr.PrePostAssign expr -> resolveLocal(expr, expr.name());
+            case Expr.PrePostIndex expr -> {
+                resolve(expr.collection());
+                resolve(expr.index());
+            }
             case Expr.PrePostSet expr -> resolve(expr.object());
             case Expr.Set expr -> {
                 resolve(expr.value());