+++ lib/src/main/java/com/wjduquette/joe/walker/Scanner.java
@@ -104,6 +104,8 @@ class Scanner {
             case ')'  -> addToken(RIGHT_PAREN);
             case '{'  -> addToken(LEFT_BRACE);
             case '}'  -> addToken(RIGHT_BRACE);
+            case '['  -> addToken(LEFT_BRACKET);
+            case ']'  -> addToken(RIGHT_BRACKET);
             case '@'  -> addToken(AT);
             case '\\' -> addToken(BACK_SLASH);
             case ':'  -> addToken(COLON);